+++ test/OpenTelemetry.Contrib.Shared.Tests/GrpcTagHelperTests.cs
@@ -3,7 +3,6 @@
 
 using System.Diagnostics;
 using OpenTelemetry.Trace;
-using Xunit;
 
 namespace OpenTelemetry.Instrumentation.Tests;
 
@@ -54,6 +53,7 @@ public class GrpcTagHelperTests
 
     [Theory]
     [InlineData(0, ActivityStatusCode.Unset)] // Ok
+    [InlineData(-1, ActivityStatusCode.Error)] // Invalid negative status code
     [InlineData(1, ActivityStatusCode.Error)] // Cancelled
     [InlineData(2, ActivityStatusCode.Error)] // Unknown
     [InlineData(3, ActivityStatusCode.Error)] // InvalidArgument
@@ -79,6 +79,7 @@ public class GrpcTagHelperTests
 
     [Theory]
     [InlineData(0, ActivityStatusCode.Unset)] // Ok
+    [InlineData(-1, ActivityStatusCode.Error)] // Invalid negative status code
     [InlineData(1, ActivityStatusCode.Unset)] // Cancelled
     [InlineData(2, ActivityStatusCode.Error)] // Unknown
     [InlineData(3, ActivityStatusCode.Unset)] // InvalidArgument