+++ test/OpenTelemetry.Exporter.Geneva.Tests/GenevaLogExporterTests.cs
@@ -15,7 +15,6 @@ using OpenTelemetry.Exporter.Geneva.MsgPack;
 using OpenTelemetry.Logs;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.Exporter.Geneva.Tests;