+++ test/OpenTelemetry.OpAmp.Client.Tests/PlainHttpTransportTests.cs
@@ -16,7 +16,6 @@ using OpenTelemetry.OpAmp.Client.Settings;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using OpenTelemetry.OpAmp.Client.Tests.Tools;
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.OpAmp.Client.Tests;
 
@@ -108,12 +107,11 @@ public class PlainHttpTransportTests
                 context.Response.OutputStream.Write(oversizedBody, 0, oversizedBody.Length);
                 context.Response.Close();
             },
-            out var host,
-            out var port);
+            out var serverUrl);
 
         var settings = new OpAmpClientSettings
         {
-            ServerUrl = new Uri($"http://{host}:{port}"),
+            ServerUrl = serverUrl,
         };
 
         var frameProcessor = new FrameProcessor();
@@ -164,12 +162,11 @@ public class PlainHttpTransportTests
                 context.Response.OutputStream.Write(compressedBody, 0, compressedBody.Length);
                 context.Response.Close();
             },
-            out var host,
-            out var port);
+            out var serverUrl);
 
         var settings = new OpAmpClientSettings
         {
-            ServerUrl = new Uri($"http://{host}:{port}"),
+            ServerUrl = serverUrl,
             HttpClientFactory = () =>
             {
                 var handler = new HttpClientHandler
@@ -234,12 +231,11 @@ public class PlainHttpTransportTests
                     thresholdReached.Set();
                 }
             },
-            out var host,
-            out var port);
+            out var serverUrl);
 
         var settings = new OpAmpClientSettings
         {
-            ServerUrl = new Uri($"http://{host}:{port}"),
+            ServerUrl = serverUrl,
         };
 
         var frameProcessor = new FrameProcessor();
@@ -307,12 +303,11 @@ public class PlainHttpTransportTests
                     }
                 }
             },
-            out var host,
-            out var port);
+            out var serverUrl);
 
         var settings = new OpAmpClientSettings
         {
-            ServerUrl = new Uri($"http://{host}:{port}"),
+            ServerUrl = serverUrl,
         };
 
         var frameProcessor = new FrameProcessor();
@@ -340,12 +335,11 @@ public class PlainHttpTransportTests
                 context.Response.OutputStream.Write(body, 0, body.Length);
                 context.Response.Close();
             },
-            out var host,
-            out var port);
+            out var serverUrl);
 
         var settings = new OpAmpClientSettings
         {
-            ServerUrl = new Uri($"http://{host}:{port}"),
+            ServerUrl = serverUrl,
         };
 
         var frameProcessor = new FrameProcessor();