+++ test/OpenTelemetry.OpAmp.Client.Tests/WsTransportTest.cs
@@ -11,7 +11,6 @@ using OpenTelemetry.OpAmp.Client.Settings;
 using OpenTelemetry.OpAmp.Client.Tests.Mocks;
 using OpenTelemetry.OpAmp.Client.Tests.Tools;
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.OpAmp.Client.Tests;
 
@@ -255,7 +254,7 @@ public class WsTransportTest
         var sendException = await Record.ExceptionAsync(
             () => wsTransport.SendAsync(FrameGenerator.GenerateMockAgentFrame().Frame, CancellationToken.None));
         Assert.True(
-            sendException is null || sendException is WebSocketException || sendException is OperationCanceledException,
+            sendException is null or WebSocketException or OperationCanceledException,
             $"Unexpected exception from SendAsync: {sendException}");
 
         Assert.True(opAmpServer.TryGetClientCloseStatus(TimeSpan.FromSeconds(5), out var closeStatus));
@@ -332,7 +331,7 @@ public class WsTransportTest
         var sendException = await Record.ExceptionAsync(
             () => wsTransport.SendAsync(FrameGenerator.GenerateMockAgentFrame().Frame, CancellationToken.None));
         Assert.True(
-            sendException is null || sendException is WebSocketException || sendException is OperationCanceledException,
+            sendException is null or WebSocketException or OperationCanceledException,
             $"Unexpected exception from SendAsync: {sendException}");
 
         Assert.True(opAmpServer.TryGetClientCloseStatus(timeout, out var closeStatus));