+++ test/OpenTelemetry.Resources.AWS.Tests/Http/ServerCertificateValidationProviderTests.cs
@@ -4,7 +4,6 @@
 #if !NETFRAMEWORK
 
 using System.Security.Cryptography.X509Certificates;
-using Xunit;
 
 namespace OpenTelemetry.Resources.AWS.Tests.Http;