+++ test/OpenTelemetry.Sampler.AWS.Tests/TestAWSXRaySamplerClient.cs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using OpenTelemetry.Tests;
-using Xunit;
 
 namespace OpenTelemetry.Sampler.AWS.Tests;
 
@@ -17,9 +16,8 @@ public class TestAWSXRaySamplerClient : IDisposable
         this.requestHandler = new MockServerRequestHandler();
         this.mockServer = TestHttpServer.RunServer(
             this.requestHandler.Handle,
-            out var host,
-            out var port);
-        this.client = new AWSXRaySamplerClient($"http://{host}:{port}");
+            out var endpoint);
+        this.client = new AWSXRaySamplerClient(endpoint.ToString().TrimEnd('/'));
     }
 
     public void Dispose()