+++ src/OpenTelemetry.Instrumentation.Remoting/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+* Updated OpenTelemetry core component version(s) to `1.16.0`.
+  ([#4487](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4487))
+
+## 0.1.0-alpha.1
+
+Released 2026-Jun-01
+
 * This is the first release of `OpenTelemetry.Instrumentation.Remoting` package.
 
 For more details, please refer to the [README](README.md).